Initial commit: Add HTML, CSS, and JS files
This commit adds the initial HTML, CSS, and JavaScript files for the
Calculator project. The HTML file contains the structure of the calculator.
It contains buttons with numbers and operators and display at the top which
shows the results.
The CSS files contains styling for the HTML.
JavaScript file contains code that will store the values of the first and
second inputted number, calculate the result and continue to calculate
whatever is inputted afterwards with that result.