refactor: Complete code replacement for calculator
New algorithm for calculation. It is now possible to enter
multiple numbers with operators with all of them still visible on screen
until '=' is clicked. When '=' is clicked result will be displayed and
expression that was calculated will be shown below it.
Add backspace button.
Add the possibility of squaring numbers.